feat: refactor module to EXM
