+++ src/modules/esm.mjs
@@ -1,0 +1,53 @@
+import * as path from 'path';
+import { fileURLToPath } from 'url';
+import { release, version } from 'os';
+import { createServer } from 'http'
+import { a } from './files/a.js'
+import { b } from './files/b.js'
+import { c } from './files/c.js';
+
+
+// const path = require('path');
+// const { release, version } = require('os');
+// const { createServer: createServerHttp } = require('http');
+// require('./files/c');
+
+const random = Math.random();
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = a;
+} else {
+    unknownObject = b;
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServer((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+c()
+
+export default {
+    unknownObject,
+    myServer,
+};
+
