+++ src/modules/files/a.js
@@ -1,4 +1,4 @@
-{
+export const a = {
     "a": 1,
     "b": 2,
     "c": 3
