+++ src/modules/files/b.js
@@ -1,4 +1,4 @@
-{
+export const b = {
     "a": 11,
     "b": 22,
     "c": 33
